Provide default run_job implementation for VM backends
